Some new features added to the admin menu

+++ Forum/Forum/Areas/Administrator/Controllers/AdministratorController.cs
@@ -1,6 +1,7 @@
 ﻿namespace Forum.Areas.Administrator.Controllers
 {
     using Forum.Data;
+    using Forum.Models.Administrator;
     using Forum.Models.Category;
     using Forum.Models.Home;
     using Forum.Services.Administrator;
@@ -24,7 +25,8 @@
         [Authorize]
         public IActionResult Manage()
         {
-            if (!User.IsInRole(GlobalConstants.Administator.AdministratorRoleName))
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
             {
                 return RedirectToAction("Error", "Home");
             }
@@ -33,7 +35,8 @@
         [Authorize]
         public IActionResult EditCategories()
         {
-            if (!User.IsInRole(GlobalConstants.Administator.AdministratorRoleName))
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
             {
                 return RedirectToAction("Error", "Home");
             }
@@ -49,7 +52,8 @@
         [Authorize]
         public IActionResult DeleteCategory(int Id)
         {
-            if (!User.IsInRole(GlobalConstants.Administator.AdministratorRoleName))
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
             {
                 return RedirectToAction("Error", "Home");
             }
@@ -69,7 +73,8 @@
         [Authorize]
         public IActionResult EditPosts()
         {
-            if (!User.IsInRole(GlobalConstants.Administator.AdministratorRoleName))
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
             {
                 return RedirectToAction("Error", "Home");
             }
@@ -85,6 +90,11 @@
         [Authorize]
         public IActionResult Category(int Id)
         {
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
+            {
+                return RedirectToAction("Error", "Home");
+            }
             var category = categoryService.GetCategory(Id);
             var postsCategory = categoryService.GetCategoryPosts(category);
             return View(new CategoryViewModel
@@ -96,12 +106,22 @@
         [Authorize]
         public IActionResult DeletePost(int Id)
         {
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
+            {
+                return RedirectToAction("Error", "Home");
+            }
             administratorService.DeletePost(Id);
             return RedirectToAction(nameof(EditPosts));
         }
         [Authorize]
         public IActionResult EditCategory(int Id)
         {
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
+            {
+                return RedirectToAction("Error", "Home");
+            }
             var categoryToEdit = data.Categories.Find(Id);
             if (categoryToEdit == null)
             {
@@ -117,12 +137,58 @@
         [HttpPost]
         public IActionResult EditCategory(int Id, EditCategoryFormModel toEdit)
         {
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
+            {
+                return RedirectToAction("Error", "Home");
+            }
             var edited = administratorService.EditCategory(Id, toEdit);
             if (!edited)
             {
                 return BadRequest();
             }
             return RedirectToAction(nameof(EditCategories));
+        }
+        [Authorize]
+        public IActionResult Users()
+        {
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
+            {
+                return RedirectToAction("Error", "Home");
+            }
+            var usersQuery = data.Users.AsQueryable();
+            var users = usersQuery
+                .OrderByDescending(x => x.Id)
+                .ToList();
+            return View(new AllUsersQuery
+            {
+                Users = users
+            });
+        }
+        [Authorize]
+        public IActionResult DeleteUser(string Id)
+        {
+            var isAdmin = IsAdmin();
+            if (!isAdmin)
+            {
+                return RedirectToAction("Error", "Home");
+            }
+            var deleted = administratorService.DeleteUser(Id);
+            if (!deleted)
+            {
+                return BadRequest();
+            }
+            return RedirectToAction(nameof(Users));
+        }
+
+        public bool IsAdmin()
+        {
+            if (User.IsInRole(GlobalConstants.Administator.AdministratorRoleName))
+            {
+                return true;
+            }
+            return false;
         }
     }
 }
